+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 3. Print errors without line breaks if not pretty printed.
 
+4. Log errors if referenced server is unknown or invalid.
+
 ### 1.4.0 (10.03.2017)
 
 1. Removed -e flag